fix: move some test in folder
